safari-534.53-branch/LayoutTests/fast/css/getComputedStyle/counterIncrement-without-counter-expected.txt

PASS style.counterIncrement is '';
PASS style.counterReset is '';
